Разбил OnboardingContainerView на компоненты, объединил Color Extensions, исправил замечания
