+++ SunaTravel/Extensions/Color+Extensions.swift
@@ -1,12 +1,13 @@
 //
-//  ColorAccentBlue.swift
+//  Color+Extensions.swift
 //  SunaTravel
 //
-//  Created by Иван Тарасюк on 13.12.2024.
+//  Created by Иван Тарасюк on 14.12.2024.
 //
 
 import SwiftUI
 
 extension Color {
+    static let deepOrange = Color(red: 1.0, green: 0.45, blue: 0.0)
     static let accentBlue = Color(red: 0x24 / 255.0, green: 0xBA / 255.0, blue: 0xEC / 255.0)
 }
