--- SunaTravel/Extensions/ColorDeepOrange.swift
@@ -1,12 +1,0 @@
-//
-//  ColorDeepOrange.swift
-//  SunaTravel
-//
-//  Created by Иван Тарасюк on 13.12.2024.
-//
-
-import SwiftUI
-
-extension Color {
-    static let deepOrange = Color(red: 1.0, green: 0.45, blue: 0.0) // Более яркий оранжевый
-}
